+++ R/goesaodc_plotScanPoints.R
@@ -22,6 +22,7 @@
 #' integer from 1 to 15.
 #' @param stateCodes Codes of state outlines to draw.
 #' @param title Title of the plot.
+#' @param legendTitle Title of the plot legend.
 #' 
 #' @examples 
 #' \donttest{
@@ -83,7 +84,8 @@ goesaodc_plotScanPoints <- function(
   includeMap = FALSE,
   zoom = NULL,
   stateCodes = NULL,
-  title = NULL
+  title = NULL,
+  legendTitle = "AOD"
 ) {
   
   # ----- Validate parameters --------------------------------------------------
@@ -112,7 +114,7 @@ goesaodc_plotScanPoints <- function(
   # Create base layer
   baseLayer <- AirFirePlots::plot_base(
     title = title,
-    clab = "AOD",
+    clab = legendTitle,
     xlim = xlim,
     ylim = ylim,
     project = includeMap,